+++ tests/common.py
@@ -139,6 +139,10 @@ a_repr = """\
 `{}`
 """
 
+complex_f_string = '''\
+f\'\'\'-{f"""*{f"+{f'.{x}.'}+"}*"""}-\'\'\'
+'''
+
 async_function_def = """\
 async def f():
     suite1
@@ -211,6 +215,7 @@ class AstunparseCommonTestCase:
         self.check_roundtrip("not True or False")
         self.check_roundtrip("True or not False")
 
+    @unittest.skipUnless(sys.version_info < (3, 6), "Only works for Python < 3.6")
     def test_integer_parens(self):
         self.check_roundtrip("3 .__abs__()")
 
@@ -316,6 +321,10 @@ class AstunparseCommonTestCase:
     @unittest.skipUnless(six.PY2, "Only for Python 2")
     def test_repr(self):
         self.check_roundtrip(a_repr)
+
+    @unittest.skipUnless(sys.version_info[:2] >= (3, 6), "Only for Python 3.6 or greater")
+    def test_complex_f_string(self):
+        self.check_roundtrip(complex_f_string)
 
     @unittest.skipUnless(six.PY3, "Only for Python 3")
     def test_annotations(self):